requirements.txt
dj_database_url==2.3.0
django-environ==0.12.0
Django==5.1.6
django-admin-interface==0.29.4
django-admin-autocomplete-filter==0.7.1
django-filter==25.1
djangorestframework==3.15.2
gunicorn==23.0.0 
psycopg2==2.9.10
redis==5.2.1
whitenoise==6.9.0